requirements.txt
rdkit-pypi
streamlit
langchain
langchain-openai
langchain-community
langchain-huggingface
chromadb
sentence-transformers
python-dotenv
openai
